pyproject.toml
[tool.poetry]
name = "foreshadwownba-dvc-ml-pipeline"
version = "0.1.0"
description = ""
authors = ["Yannick Flores <yannick.flores1992@gmail.com>"]
readme = "README.md"
packages = [{include = "foreshadwownba_dvc_ml_pipeline"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
ipykernel = "6.29.*"
pandas = "2.0.0"
numpy = "1.26.*"
dvc = {extras = ["s3"], version = "^3.38.1"}
pymysql = "1.1.1"
sqlalchemy = "2.0.*"
python-dotenv  = "1.0.*"
scikit-learn = "1.4.*"
python-decouple = "3.6"
pycaret = "3.3"
kaleido = "0.2.1"
joblib = "1.3.2"
shap = "*"
xgboost = "*"
auto-feat-ml = "*"
mlxtend = "0.23.*"
feature_engine = "1.6.*"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "6.5.*"}
black = {version = "22.10.*", allow-prereleases = true}
pytest = "7.2.*"
pytest-html = "3.1.*"
pytest-cov = "2.12.*"
py = "1.11.*"
pylint = "2.17.*"
python-decouple = "3.6"
mkdocs = "1.4.*"
importlib-resources="6.0.1"
pandantic = "0.3.*"
# skimpy = "0.0.*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
addopts = [
    "--import-mode=importlib",
]
